requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
google-cloud-vision>=2.4.2
httpx>=0.23.0
pillow>=8.3.1
opencv-python>=4.5.3
python-dotenv>=0.19.0
pydantic>=1.8.2